Polish 3D globe shaders and split network animations
- Per-dot twinkle via InstancedBufferAttribute + onBeforeCompile
- Aurora-rippling atmosphere material; outer cinematic halo pass
- Hard-hide sphere meshes when shellProgress is ~0 to fix the
  black-disc artifact during the globe→flat transition
- Solid render mode in flat view falls back to a worldFill-colored
  dot field (was rendering empty)
- Split the live network into independent Arcs / Pulses sub-toggles
  with a strength slider; defaults: network on, both on, strength 70
- Space mesh fragment shader uses mediump float for perf; post-effect
  pass retains highp
- WebGL context loss / restored handlers wired up

+++ src/three/space-mesh.js
@@ -16,7 +16,12 @@ const VERTEX_SHADER = /* glsl */ `
 `;
 
 const FRAGMENT_SHADER = /* glsl */ `
-  precision highp float;
+  // mediump is sufficient — the math operates on 0–1 UV ranges, small-magnitude
+  // hash outputs, and bounded color values. Dropping from highp gives a real
+  // perf win on mobile/integrated GPUs (MDN WebGL best-practices recommendation)
+  // while being visually identical on desktop hardware that treats mediump
+  // as highp anyway.
+  precision mediump float;
 
   varying vec2 vUv;
   uniform vec2 uResolution;
